simplificando algumas configurações
