+++ src/main/java/br/com/academiadev/thunderpets/config/SwaggerConfig.java
@@ -1,34 +1,20 @@
 package br.com.academiadev.thunderpets.config;
 
-import org.springframework.beans.factory.annotation.Value;
 import org.springframework.context.annotation.Bean;
 import org.springframework.context.annotation.Configuration;
 import springfox.documentation.builders.PathSelectors;
 import springfox.documentation.builders.RequestHandlerSelectors;
-import springfox.documentation.service.*;
+import springfox.documentation.service.ApiInfo;
+import springfox.documentation.service.Contact;
 import springfox.documentation.spi.DocumentationType;
-import springfox.documentation.spi.service.contexts.SecurityContext;
 import springfox.documentation.spring.web.plugins.Docket;
-import springfox.documentation.swagger.web.ApiKeyVehicle;
-import springfox.documentation.swagger.web.SecurityConfiguration;
 import springfox.documentation.swagger2.annotations.EnableSwagger2;
 
 import java.util.ArrayList;
-import java.util.Collections;
-import java.util.List;
 
 @Configuration
 @EnableSwagger2
 public class SwaggerConfig {
-
-    @Value("${oauth.client}")
-    private String client;
-
-    @Value("${oauth.secret}")
-    private String secret;
-
-    @Value("${oauth.url}")
-    private String url;
 
     @Bean
     public Docket apis() {
@@ -36,9 +22,7 @@ public class SwaggerConfig {
                 .apis(RequestHandlerSelectors.basePackage("br.com.academiadev.thunderpets"))
                 .paths(PathSelectors.any())
                 .build()
-                .apiInfo(apiInfo())
-                .securitySchemes(Collections.singletonList(securitySchema()))
-                .securityContexts(Collections.singletonList(securityContext()));
+                .apiInfo(apiInfo());
     }
 
     private ApiInfo apiInfo() {
@@ -53,41 +37,5 @@ public class SwaggerConfig {
                 "License",
                 "MIT",
                 new ArrayList<>());
-    }
-
-    private OAuth securitySchema() {
-        List<AuthorizationScope> authorizationScopeList = new ArrayList<>();
-
-        authorizationScopeList.add(new AuthorizationScope("read", "read all"));
-        authorizationScopeList.add(new AuthorizationScope("trust", "trust all"));
-        authorizationScopeList.add(new AuthorizationScope("write", "access all"));
-
-        List<GrantType> grantTypes = new ArrayList<>();
-        GrantType creGrant = new ResourceOwnerPasswordCredentialsGrant(url);
-        grantTypes.add(creGrant);
-
-        return new OAuth("oauth2schema", authorizationScopeList, grantTypes);
-    }
-
-    private SecurityContext securityContext() {
-        return SecurityContext.builder()
-                .securityReferences(defaultAuth())
-                .forPaths(PathSelectors.ant("/pet/**"))
-                .build();
-    }
-
-    private List<SecurityReference> defaultAuth() {
-        final AuthorizationScope[] authorizationScopes = new AuthorizationScope[3];
-
-        authorizationScopes[0] = new AuthorizationScope("read", "read all");
-        authorizationScopes[1] = new AuthorizationScope("trust", "trust all");
-        authorizationScopes[2] = new AuthorizationScope("write", "write all");
-
-        return Collections.singletonList(new SecurityReference("oauth2schema", authorizationScopes));
-    }
-
-    @Bean
-    public SecurityConfiguration securityInfo() {
-        return new SecurityConfiguration(client, secret, "", "", "", ApiKeyVehicle.HEADER, "", " ");
     }
 }
