+++ src/main/java/br/com/academiadev/thunderpets/config/security/AuthorizationServerConfig.java
@@ -11,7 +11,6 @@ import org.springframework.security.oauth2.config.annotation.web.configuration.A
 import org.springframework.security.oauth2.config.annotation.web.configuration.EnableAuthorizationServer;
 import org.springframework.security.oauth2.config.annotation.web.configurers.AuthorizationServerEndpointsConfigurer;
 import org.springframework.security.oauth2.config.annotation.web.configurers.AuthorizationServerSecurityConfigurer;
-import org.springframework.security.web.AuthenticationEntryPoint;
 
 @Configuration
 @EnableAuthorizationServer
@@ -20,10 +19,10 @@ public class AuthorizationServerConfig extends AuthorizationServerConfigurerAdap
     private static final Integer ACCESS_TOKEN_VALIDITY_IN_SECONDS = 60 * 60 * 24;
     private static final Integer REFRESH_TOKEN_VALIDITY_IN_SECONDS = 60 * 60 * 24;
 
-    @Value("${oauth.client}")
+    @Value("${security.oauth2.client.client-id}")
     private String client;
 
-    @Value("${oauth.secret}")
+    @Value("${security.oauth2.client.client-secret}")
     private String secret;
 
     @Autowired
