+++ src/main/java/br/com/academiadev/thunderpets/config/security/ResourceServerConfig.java
@@ -1,6 +1,7 @@
 package br.com.academiadev.thunderpets.config.security;
 
 import org.springframework.context.annotation.Configuration;
+import org.springframework.http.HttpMethod;
 import org.springframework.security.config.annotation.web.builders.HttpSecurity;
 import org.springframework.security.oauth2.config.annotation.web.configuration.EnableResourceServer;
 import org.springframework.security.oauth2.config.annotation.web.configuration.ResourceServerConfigurerAdapter;
@@ -12,6 +13,6 @@ public class ResourceServerConfig extends ResourceServerConfigurerAdapter {
     @Override
     public void configure(HttpSecurity http) throws Exception {
         http.authorizeRequests()
-                .antMatchers("/pet**/**").authenticated();
+                .antMatchers(HttpMethod.POST, "/pet**/**").authenticated();
     }
 }
