+++ src/main/java/br/com/academiadev/thunderpets/config/security/WebConfigSecurityAdapter.java
@@ -1,6 +1,5 @@
 package br.com.academiadev.thunderpets.config.security;
 
-import br.com.academiadev.thunderpets.dto.UsuarioCustomDTO;
 import br.com.academiadev.thunderpets.model.Usuario;
 import br.com.academiadev.thunderpets.repository.UsuarioRepository;
 import org.springframework.beans.factory.annotation.Autowired;
@@ -8,6 +7,7 @@ import org.springframework.context.annotation.Bean;
 import org.springframework.context.annotation.Configuration;
 import org.springframework.security.authentication.AuthenticationManager;
 import org.springframework.security.config.annotation.authentication.builders.AuthenticationManagerBuilder;
+import org.springframework.security.config.annotation.web.builders.WebSecurity;
 import org.springframework.security.config.annotation.web.configuration.EnableWebSecurity;
 import org.springframework.security.config.annotation.web.configuration.WebSecurityConfigurerAdapter;
 import org.springframework.security.crypto.bcrypt.BCryptPasswordEncoder;
@@ -17,7 +17,7 @@ import org.springframework.security.crypto.bcrypt.BCryptPasswordEncoder;
 public class WebConfigSecurityAdapter extends WebSecurityConfigurerAdapter {
 
     @Bean
-    public static BCryptPasswordEncoder passwordEncoder() {
+    private static BCryptPasswordEncoder passwordEncoder() {
         return new BCryptPasswordEncoder();
     }
 
@@ -39,7 +39,17 @@ public class WebConfigSecurityAdapter extends WebSecurityConfigurerAdapter {
             usuarioRepository.save(usuario);
         }
 
-        builder.userDetailsService(email -> new UsuarioCustomDTO(usuarioRepository.findOneByEmail(email)))
-                .passwordEncoder(passwordEncoder());
+        builder.userDetailsService(usuarioRepository::findOneByEmail).passwordEncoder(passwordEncoder());
+    }
+
+    @Override
+    public void configure(WebSecurity web) throws Exception {
+        web.ignoring().antMatchers(
+                "/v2/api-docs",
+                "/configuration/ui",
+                "/swagger-resources",
+                "/configuration/security",
+                "/swagger-ui.html",
+                "/webjars/**");
     }
 }
