+++ src/main/resources/application.properties
@@ -13,6 +13,6 @@ spring.jpa.generate-ddl=true
 spring.jpa.show-sql=true
 spring.jpa.format-sql=true
 
-oauth.client=${OAUTH_CLIENT:client-id}
-oauth.secret=${OAUTH_SECRET:secret-id}
 oauth.url=${OAUTH_URL:http://localhost:8080}/oauth/token
+security.oauth2.client.client-id=${OAUTH_CLIENT:client-id}
+security.oauth2.client.client-secret=${OAUTH_SECRET:secret-id}
